+++ src/onemail/om-lambda-sender/src/repositories/metrics.repository.ts
@@ -5,7 +5,7 @@ import {
   type MetricDatum,
   PutMetricDataCommand,
 } from '@aws-sdk/client-cloudwatch';
-import { isEmpty } from 'lodash';
+import isEmpty from 'lodash-es/isEmpty.js';
 
 export enum SenderMetricName {
   EmailBatchNotFound = 'EmailBatchNotFound',