+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # onemail
 
+## 1.1.19
+
+### Patch Changes
+
+- Updated dependencies [6af8fec]
+  - om-ecs-dispatcher@1.5.5
+  - om-lambda-sender@1.2.2
+
+## 1.1.18
+
+### Patch Changes
+
+- Updated dependencies [4e6a558]
+  - infra-domains-onemail-app@1.7.0
+
 ## 1.1.17
 
 ### Patch Changes